refactor: extract storefront helper orchestration
